Rename `ConnectionBuilder` to `ConnectionSpec`
Currently the type that is used to specify the parameters to create a
connection is called `ConnectionBuilder`. This name suggests that it is
capable of doing something, of building a connection, but it actually
doesn't have any method, it is just a container for the parameters. To
make that explicit is better to avoid the action oriented name,
`ConnectionSpec` is probably better.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>