Remove unneeded `vendor` link
Currently the `build.py` script creates a link to the `vendor` directory
inside the Go source directory, because this was needed in a previous
incarnation of the script that was used to build an application where it
was necessary to generate Kubernetes client code. This is no longer
needed, so this patch removes it.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>